Complete blackjack game code
Completed the whole blackjack code It was so much fun. In this process with the help of Angela Yu's hint I created a lot of functions, such as deal_card, play_game(), compare() , With those I also learned how can I have this all in great structure so It will be running smoothly.